requirements.txt
asgiref==3.3.4
Django==3.2.3
django-unused-media==0.2.2
djangorestframework==3.12.4
drf-extensions==0.7.0
Faker==8.4.0
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
